Add contact info next to the contact message
