+++ app/components/contact_info_component.html.erb
@@ -1,0 +1,20 @@
+  <% if has_contact_information? %>
+    <div class="row">
+      <% if phone.present? %>
+        <div class="col-sm-4 mb-2">
+          <%= icon('phone') { PhoneNumber.new(phone).to_s } %>
+        </div>
+        <div class="col-sm-4 mb-2">
+          <%= icon('whatsapp') { whatsapp_api_link  } %>
+        </div>
+        <div class="col-sm-4 mb-2">
+          <%= icon('whatsapp') { whatsapp_web_link  } %>
+        </div>
+      <% end %>
+      <div class="col-sm-12 mb-2">
+        <% if email.present? %>
+          <%= icon('pencil') { email } %>
+        <% end %>
+      </div>
+    </div>
+  <% end %>
