+++ app/components/public_talks/contact_speaker_message_component.html.erb
@@ -1,3 +1,5 @@
+<%= render contact_info %>
+
 <%= render CopyToClipboardComponent.new(selector: '#contact-message-text') %>
 
 <div id="contact-message-container">
