+++ app/components/public_talks/speakers/show_page_component.rb
@@ -21,33 +21,9 @@ class PublicTalks::Speakers::ShowPageComponent < PageComponent
     )
   end
 
-  def has_contact_information?
-    [speaker.email, speaker.phone].map(&:presence).compact.any?
-  end
-
-  def whatsapp_web_link
-    link_to(
-      'Web',
-      "https://web.whatsapp.com/send?phone=#{phone_number}",
-      target: :blank
-    )
-  end
-
-  def whatsapp_api_link
-    link_to(
-      'APP',
-      "https://api.whatsapp.com/send?phone=#{phone_number}",
-      target: :blank
-    )
-  end
-
   private
 
   def delete_confirmation(speaker)
     t('app.messages.confirm_delete_x', record: speaker.name)
-  end
-
-  def phone_number
-    "55#{PhoneNumber.new(speaker.phone).unformatted}"
   end
 end
